scraper/requirements.txt
facebook_scraper
python-dotenv==1.0.0
selenium==4.10.0
